+++ cloudfeaster/spider.py
@@ -356,8 +356,6 @@ class CrawlResponseOk(CrawlResponse):
     def __init__(self, *args, **kwargs):
         kwargs['status_code'] = CrawlResponse.SC_OK
         kwargs['status'] = 'Ok'
-        if 1 == len(args):
-            kwargs['data'] = args[0]
         CrawlResponse.__init__(self, *args, **kwargs)